Rename SpecialFunction to EndpointFunction
