+++ src/functions/GeneralFunction.java
@@ -7,8 +7,8 @@ import functions.binary.integer.IntegerBinaryFunction;
 import functions.commutative.Product;
 import functions.commutative.Sum;
 import functions.commutative.integer.IntegerCommutativeFunction;
-import functions.special.Constant;
-import functions.special.Variable;
+import functions.endpoint.Constant;
+import functions.endpoint.Variable;
 import functions.unitary.integer.IntegerUnitaryFunction;
 import functions.unitary.piecewise.PiecewiseFunction;
 import functions.unitary.specialcases.SpecialCaseBinaryFunction;
