+++ src/functions/binary/Logb.java
@@ -3,7 +3,7 @@ package functions.binary;
 import functions.GeneralFunction;
 import functions.commutative.Product;
 import functions.commutative.Sum;
-import functions.special.Constant;
+import functions.endpoint.Constant;
 import functions.unitary.specialcases.Exp;
 import functions.unitary.specialcases.Ln;
 import output.OutputBinary;
