+++ src/functions/commutative/Sum.java
@@ -2,7 +2,7 @@ package functions.commutative;
 
 import functions.GeneralFunction;
 import functions.Outputable;
-import functions.special.Constant;
+import functions.endpoint.Constant;
 import output.OutputCommutative;
 import output.OutputFunction;
 import tools.ArrayTools;
