+++ src/functions/unitary/trig/inverse/Asech.java
@@ -5,7 +5,7 @@ import functions.GeneralFunction;
 import functions.binary.Pow;
 import functions.commutative.Sum;
 import functions.commutative.Product;
-import functions.special.Constant;
+import functions.endpoint.Constant;
 import functions.unitary.UnitaryFunction;
 import functions.unitary.piecewise.Abs;
 import functions.unitary.trig.normal.Sech;
