+++ src/tools/DefaultFunctions.java
@@ -5,8 +5,8 @@ import functions.binary.Logb;
 import functions.binary.Pow;
 import functions.commutative.Product;
 import functions.commutative.Sum;
-import functions.special.Constant;
-import functions.special.Variable;
+import functions.endpoint.Constant;
+import functions.endpoint.Variable;
 import functions.unitary.integer.combo.Factorial;
 
 /**
