+++ src/tools/integration/StageOne.java
@@ -6,8 +6,8 @@ import functions.binary.Pow;
 import functions.commutative.CommutativeFunction;
 import functions.commutative.Product;
 import functions.commutative.Sum;
-import functions.special.Constant;
-import functions.special.Variable;
+import functions.endpoint.Constant;
+import functions.endpoint.Variable;
 import functions.unitary.specialcases.Exp;
 import functions.unitary.specialcases.Ln;
 import functions.unitary.transforms.Integral;
