+++ src/tools/singlevariable/Solver.java
@@ -4,7 +4,7 @@ import config.Settings;
 import functions.GeneralFunction;
 import functions.binary.Pow;
 import functions.commutative.Product;
-import functions.special.Constant;
+import functions.endpoint.Constant;
 import functions.unitary.piecewise.Abs;
 import tools.DefaultFunctions;
 import tools.SolverTools;
