+++ src/tools/singlevariable/TaylorSeries.java
@@ -5,8 +5,8 @@ import functions.GeneralFunction;
 import functions.binary.Pow;
 import functions.commutative.Product;
 import functions.commutative.Sum;
-import functions.special.Constant;
-import functions.special.Variable;
+import functions.endpoint.Constant;
+import functions.endpoint.Variable;
 import tools.MiscTools;
 import tools.VariableTools;
