+++ tests/KeywordInterfaceTest.java
@@ -3,8 +3,8 @@ import functions.GeneralFunction;
 import functions.binary.Pow;
 import functions.commutative.Product;
 import functions.commutative.Sum;
-import functions.special.Constant;
-import functions.special.Variable;
+import functions.endpoint.Constant;
+import functions.endpoint.Variable;
 import functions.unitary.trig.normal.Cos;
 import functions.unitary.trig.normal.Sin;
 import org.junit.jupiter.api.Test;
